feat: paste recipe with pic to Batman cookies page
The Batman cookies page has received an appetizing photo of the cookies
(including alternative text for better accessibility),
a describing paragraph, a list of ingredients and of steps to follow
along, as well as a closing paragraph.

The paragraphs are each divided by headings for better readability (apart
from the closing paragraph which can do without).