feat: add about section and pics to homepage
A short "About" section outlines the purpose of this website
and introduces the contributers. Images of the contributers
make the experience more personal and therefore engaging for
the user.

The paragraphs are divided by different headings to create a
clear structure.

Images were added to the recipe links as a little preview and
to create more visual variety.

Alternative texts for the images were included for better
accessibility.