feat: style the header of all pages
Added background color to the header and changed the text color accordingly. Aligned the banner image with the
background and made it more responsive. This was done to create the appearance of an actual banner for better
branding and recognition. Added another icon for better balance.

The header was added to all pages for better orientation.

Adjusted the code of the recipe-table to more specificity since more tables have been and will be added to the
homepage.

A table for the header content was added for better spacing.