feat: adjust header to match branding and add links
Changed the background color of the heading to match the brand colors better and added a gradient for more
sophisticated softness. Also added links back to the homepage to every element of the header. Applied the changes
to all pages. Removed the underline on the heading link for better visual appeal.