feat: add and style footer on all pages
Added footer with background color to include the disclaimers. Added a copyright disclaimer. Adjusted margin and
padding of the footer to match the header. This distinguishes the disclaimers from the actual content of the site
and ties the page with the branding together.

Added a bit more space between the last two paragraphs on the recipe pages for better visual structure.