feat: exchange recipe table with recipe cards on homepage
Provided more visual appeal and information for each recipe on the homepage, using the box model and different flows.
Also fixed the issue with the recipe table not being responsive to narrow browser windows.
Added an additional, explicit button that links to the respective recipe page, in addition to the link in the title and picture.