feat: exchange header for navigation bar on homepage
The table in the header was removed in favor of a navigation bar that is responsive to different sizes of the browser window.
The logo is positioned left while a menu was added to the right. It includes a dropdown menu to reach the individual recipe pages.
The buttons are now responsive to hovering over them with the cursor.